refactor(competence dialog): move to separate component)
